move Consent classes to their own files
