+++ cmplibrary/src/main/java/com/sourcepoint/cmplibrary/CustomPurposeConsent.java
@@ -1,0 +1,7 @@
+package com.sourcepoint.cmplibrary;
+
+public class CustomPurposeConsent extends Consent {
+    CustomPurposeConsent(String id, String name) {
+        super(id, name, "CustomPurposeConsent");
+    }
+}
